# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meticulous API",
    "description": "REST API for the Meticulous CI/CD platform",
    "contact": {
      "name": "Meticulous Team"
    },
    "license": {
      "name": "UNLICENSED"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/v1/pipelines/{pipeline_id}/triggers": {
      "get": {
        "tags": [
          "triggers"
        ],
        "operationId": "list_triggers",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "description": "Pipeline ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Triggers for pipeline",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TriggerPublicResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "triggers"
        ],
        "operationId": "create_trigger",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "description": "Pipeline ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Trigger created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerPublicResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/triggers/{trigger_id}": {
      "delete": {
        "tags": [
          "triggers"
        ],
        "operationId": "delete_trigger",
        "parameters": [
          {
            "name": "trigger_id",
            "in": "path",
            "description": "Trigger ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trigger deleted"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "triggers"
        ],
        "operationId": "update_trigger",
        "parameters": [
          {
            "name": "trigger_id",
            "in": "path",
            "description": "Trigger ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Trigger updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerPublicResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTriggerRequest": {
        "type": "object",
        "required": [
          "kind",
          "config"
        ],
        "properties": {
          "config": {
            "type": "object"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "generate_webhook_secret": {
            "type": "boolean",
            "description": "When true and kind is `webhook`, a random secret is generated and returned once in the response."
          },
          "kind": {
            "type": "string"
          }
        }
      },
...
      "TriggerPublicResponse": {
        "type": "object",
        "required": [
          "id",
          "pipeline_id",
          "kind",
          "config",
          "secret_configured",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "config": {
            "type": "object",
            "description": "Webhook configuration without `secret`; use `secret_configured` to see if HMAC is required."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "generated_secret": {
            "type": [
              "string",
              "null"
            ],
            "description": "Present only on create when `generate_webhook_secret` was true."
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "pipeline_id": {
            "type": "string"
          },
          "secret_configured": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateTriggerRequest": {
        "type": "object",
        "required": [
          "config_patch"
        ],
        "properties": {
          "config_patch": {
            "type": "object",
            "description": "Merged into existing JSON config (objects merge recursively)."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}